cleanup args and types
